[oracles] move oracles logic to its own module
